Fix Memoizer.getMemoFile(id) when directory is the root path

+++ components/formats-bsd/test/loci/formats/utests/MemoizerTest.java
@@ -39,6 +39,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.UUID;
 
 import loci.common.Location;
 import loci.formats.FormatException;
@@ -101,7 +102,7 @@ public class MemoizerTest {
   public void testSimple() throws Exception {
       memoizer = new Memoizer(reader);
       File f = memoizer.getMemoFile(TEST_FILE);
-      if (f.exists()) {
+      if (f != null && f.exists()) {
         f.delete();
       }
       // At this point we're sure that there's no memo file.
@@ -111,6 +112,53 @@ public class MemoizerTest {
       memoizer.close();
       memoizer.setId(TEST_FILE);
       memoizer.close();
+  }
+
+  @Test
+  public void testGetMemoFileNoDirectory() throws Exception {
+      memoizer = new Memoizer(reader);
+      File f = memoizer.getMemoFile(TEST_FILE);
+      assertEquals(f, null);
+  }
+
+  @Test
+  public void testGetMemoFileNoDirectory2() throws Exception {
+      memoizer = new Memoizer(reader, 0);
+      File f = memoizer.getMemoFile(TEST_FILE);
+      assertEquals(f, null);
+  }
+
+  @Test
+  public void testGetMemoFileDirectory() throws Exception {
+      File directory = new File(System.getProperty("java.io.tmpdir"));
+      memoizer = new Memoizer(reader, 0, directory);
+      String idDir = System.getProperty("java.io.tmpdir");
+      String id = new File(idDir, TEST_FILE).getAbsolutePath();
+      String memoDir = idDir.substring(idDir.indexOf(File.separator) + 1);
+      File memoFile = new File(directory, memoDir);
+      memoFile = new File(memoFile, "." + TEST_FILE + ".bfmemo");
+      assertEquals(memoizer.getMemoFile(id).getAbsolutePath(),
+                   memoFile.getAbsolutePath());
+  }
+
+  @Test
+  public void testGetMemoFileNonExistingDirectory() throws Exception {
+      String uuid = UUID.randomUUID().toString();
+      File directory = new File(System.getProperty("java.io.tmpdir") + uuid);
+      memoizer = new Memoizer(reader, 0, directory);
+      File f = memoizer.getMemoFile(TEST_FILE);
+      assertEquals(f, null);
+  }
+
+  @Test
+  public void testGetMemoFileSameDirectory() throws Exception {
+      String idDir = System.getProperty("java.io.tmpdir");
+      String id = new File(idDir, TEST_FILE).getAbsolutePath();
+      String rootPath = idDir.substring(0, idDir.indexOf(File.separator) + 1);
+      memoizer = new Memoizer(reader, 0, new File(rootPath));
+      File memoFile = new File(idDir, "." + TEST_FILE + ".bfmemo");
+      assertEquals(memoizer.getMemoFile(id).getAbsolutePath(),
+                   memoFile.getAbsolutePath());
   }
 
   public static void main(String[] args) throws Exception {
